Remove "explore" components folder
