--- apps/app/src/constants/colors.ts
@@ -1,30 +1,0 @@
-export const light = {
-	text: {},
-	views: {},
-	buttons: {}
-};
-
-export const dark = {
-	text: {},
-	views: {},
-	buttons: {}
-};
-
-// Experiment: Just use iOS human interface guidelines colors, to save time.
-export const newLight = {
-	gray: 'rgb(142, 142, 147)',
-	gray2: 'rgb(174, 174, 178)',
-	gray3: 'rgb(119, 119, 204)',
-	gray4: 'rgb(209, 209, 214)',
-	gray5: 'rgb(229, 229, 234)',
-	gray6: 'rgb(242, 242, 247)'
-};
-
-export const newDark = {
-	gray: 'rgb(142, 142, 147)',
-	gray2: 'rgb(99, 99, 102)',
-	gray3: 'rgb(72, 72, 74)',
-	gray4: 'rgb(58, 58, 60)',
-	gray5: 'rgb(44, 44, 46)',
-	gray6: 'rgb(28, 28, 30)'
-};
