--- apps/app/src/constants/fonts.ts
@@ -1,29 +1,0 @@
-// Remember:
-// * This is lifted from the iOS Human Interface Guidelines: https://developer.apple.com/design/human-interface-guidelines/ios/visual-design/typography
-// * Leading -> Line Height.
-// * Make sure that these work well with Android.
-
-export const sizes = {
-	// Large Title (Weight: Regular, Leading: 41pt).
-	largeTitle: 34,
-	// Title 1 (Weight: Regular, Leading 34pt)
-	title1: 28,
-	// Title 2 (Weight: Regular, Leading 28pt)
-	title2: 22,
-	// Title 3 (Weight: Regular, Leading 25pt)
-	title3: 20,
-	// Headline (Weight: Semi-Bold, Leading 22pt)
-	headline: 17,
-	// Body (Weight: Regular, Leading 22pt)
-	body: 17,
-	// Callout (Weight: Regular, Leading 21pt)
-	callout: 16,
-	// Subhead (Weight: Regular, Leading 20pt)
-	subhead: 15,
-	// Footnote (Weight: Regular, Leading 18pt)
-	footnote: 13,
-	// Caption 1 (Weight: Regular, Leading 16pt)
-	caption1: 12,
-	// Caption 2 (Weight: Regular, Leading 13pt)
-	caption2: 11
-};
